test(updateCardScreenPosition): add unit tests
